reduce min time between commands
in the official tello python code they send the next command immediately after
receiving a response to the last one.

With a framerate of 120 in the pygame example i experienced no noticable input
lag when controlling the drone.
This commit also includes some other tidy ups to the pygame example.